+++ packages/server/src/registry/index.ts
@@ -4,5 +4,13 @@
 export {
   createFlowRegistry,
   InMemoryFlowRegistry,
-  type FlowRegistry
+  type FlowRegistry,
+  type SharedSchemasDescription,
+  type SharedScopeDescription
 } from "./flow-registry";
+
+export {
+  CrossFlowSchemaConflictError,
+  type ConflictScope,
+  type CrossFlowSchemaConflictDetails
+} from "./errors";